# insomnia/openapi.yaml
openapi: 3.0.3
info:
  title: Neon Life Drawing API
  description: |
    API for managing life drawing classes, models, venues, and calendar bookings.
    
    ## Authentication
    Most endpoints require a JWT Bearer token. Obtain one via the `/api/auth/login` or `/api/auth/register` endpoints.
    
    ## Base URLs
    - Production: `https://neon-api-production.fliptopbox.workers.dev`
    - Local: `http://localhost:8787`
  version: 1.0.0
  contact:
    name: Neon API Support
servers:
  - url: https://neon-api-production.fliptopbox.workers.dev
    description: Production server
  - url: http://localhost:8787
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        emailaddress:
          type: string
        active:
          type: integer
        created_on:
          type: string
          format: date-time
        login_on:
          type: string
          format: date-time
        is_admin:
          type: boolean
        fullname:
          type: string
        known_as:
          type: string
        description:
          type: string
...
    Model:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        sex:
          type: integer
          description: "0: Not specified, 1: Male, 2: Female, 3: Non-binary"
        instagram:
          type: string
        portrait:
          type: string
          description: URL to portrait image
        account_holder:
          type: string
        account_number:
          type: string
        account_sortcode:
          type: string
        active:
          type: integer
        created_on:
          type: string
          format: date-time
        modified_on:
          type: string
          format: date-time
        fullname:
          type: string
        known_as:
          type: string
        description:
          type: string
        bio_instagram:
          type: string
        websites:
          type: array
          items:
            type: string
        phone:
          type: string
        email:
          type: string
    Artist:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        active:
          type: integer
        created_on:
          type: string
          format: date-time
        modified_on:
          type: string
          format: date-time
        fullname:
          type: string
        known_as:
          type: string
        description:
          type: string
        instagram:
          type: string
        websites:
          type: array
          items:
            type: string
    Venue:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        week_day:
          type: integer
          description: "0: Sunday, 1: Monday, ..., 6: Saturday"
        frequency:
          type: string
          enum: [weekly, biweekly, monthly]
        instagram:
          type: string
        website:
          type: string
        address:
          type: string
        timezone:
          type: string
        start_time:
          type: string
          description: Time in HH:MM format
        duration:
          type: integer
          description: Duration in minutes
        postcode:
          type: string
        area:
          type: string
        price_inperson:
          type: integer
          description: Price in pence
        price_online:
          type: integer
          description: Price in pence
        active:
          type: integer
        tags:
          type: array
          items:
            type: string
        created_on:
          type: string
          format: date-time
        modified_on:
          type: string
          format: date-time
        organizer_name:
          type: string
    CalendarEvent:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        venue_id:
          type: integer
        date:
          type: string
          format: date
        attendance_inperson:
          type: integer
        attendance_online:
          type: integer
        start:
          type: string
          description: Start time
        duration:
          type: integer
          description: Duration in minutes
        notes:
          type: string
        tbc:
          type: integer
          description: "1 if to be confirmed"
        fullname:
          type: string
          description: Model name
        venue_name:
          type: string
    VenueTag:
      type: object
      properties:
        id:
          type: string
          description: Lowercase tag identifier
        description:
          type: string
    Image:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        type_id:
          type: integer
        src:
          type: string
          description: Image URL
        active:
          type: integer
        created_on:
          type: string
          format: date-time
        modified_on:
          type: string
          format: date-time
        type_name:
          type: string
...
paths:
...
  /api/users:
    get:
      tags:
        - Users
      summary: List all users (Admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required
...
  /api/models:
    get:
      tags:
        - Models
      summary: List all active models
      responses:
        '200':
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Model'
    post:
      tags:
        - Models
      summary: Create or update model profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullname:
                  type: string
                known_as:
                  type: string
                description:
                  type: string
                bio_instagram:
                  type: string
                websites:
                  type: string
                  description: Newline-separated URLs
                sex:
                  type: integer
                instagram:
                  type: string
                portrait:
                  type: string
                account_holder:
                  type: string
                account_number:
                  type: string
                account_sortcode:
                  type: string
      responses:
        '201':
          description: Model created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '401':
          description: Unauthorized
...
  /api/artists:
    get:
      tags:
        - Artists
      summary: List all active artists
      responses:
        '200':
          description: List of artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
    post:
      tags:
        - Artists
      summary: Create artist profile
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Artist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '401':
          description: Unauthorized
...
  /api/venues:
    get:
      tags:
        - Venues
      summary: List all active venues
      responses:
        '200':
          description: List of venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Venue'
    post:
      tags:
        - Venues
      summary: Create venue
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - week_day
                - start_time
                - duration
              properties:
                name:
                  type: string
                week_day:
                  type: integer
                frequency:
                  type: string
                  default: weekly
                instagram:
                  type: string
                website:
                  type: string
                address:
                  type: string
                timezone:
                  type: string
                start_time:
                  type: string
                duration:
                  type: integer
                postcode:
                  type: string
                area:
                  type: string
                price_inperson:
                  type: integer
                price_online:
                  type: integer
                tags:
                  type: array
                  items:
                    type: string
            example:
              name: Life Drawing at The Studio
              week_day: 3
              frequency: weekly
              start_time: "19:00"
              duration: 180
              address: 123 Art Street
              postcode: E1 1AA
              area: East London
              price_inperson: 1500
              price_online: 1000
              tags: ["clothed", "portrait"]
      responses:
        '201':
          description: Venue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'
        '401':
          description: Unauthorized
...
  /api/calendar:
    get:
      tags:
        - Calendar
      summary: List all calendar events
      responses:
        '200':
          description: List of calendar events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CalendarEvent'
    post:
      tags:
        - Calendar
      summary: Create calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - venue_id
                - date
                - attendance_inperson
                - attendance_online
                - start
                - duration
              properties:
                user_id:
                  type: integer
                venue_id:
                  type: integer
                date:
                  type: string
                  format: date
                attendance_inperson:
                  type: integer
                attendance_online:
                  type: integer
                start:
                  type: string
                duration:
                  type: integer
                notes:
                  type: string
                tbc:
                  type: integer
            example:
              user_id: 1
              venue_id: 5
              date: "2025-01-15"
              attendance_inperson: 20
              attendance_online: 5
              start: "19:00"
              duration: 180
              notes: "Figure drawing session"
              tbc: 0
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: integer
        '400':
          description: Missing required fields or invalid references
  /api/calendar/{id}:
    get:
      tags:
        - Calendar
      summary: Get calendar event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Calendar event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEvent'
        '404':
          description: Event not found
    put:
      tags:
        - Calendar
      summary: Update calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - venue_id
                - date
                - attendance_inperson
                - attendance_online
                - start
                - duration
              properties:
                user_id:
                  type: integer
                venue_id:
                  type: integer
                date:
                  type: string
                  format: date
                attendance_inperson:
                  type: integer
                attendance_online:
                  type: integer
                start:
                  type: string
                duration:
                  type: integer
                notes:
                  type: string
                tbc:
                  type: integer
      responses:
        '200':
          description: Event updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Missing required fields or invalid references
        '404':
          description: Event not found
    delete:
      tags:
        - Calendar
      summary: Delete calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Event deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Event not found

  # Dashboard endpoints
...
  /api/venue-tags:
    get:
      tags:
        - Venue Tags
      summary: List all venue tags
      responses:
        '200':
          description: List of venue tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VenueTag'
    post:
      tags:
        - Venue Tags
      summary: Create venue tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - description
              properties:
                id:
                  type: string
                  pattern: '^[a-z]+$'
                  description: Lowercase letters only
                description:
                  type: string
            example:
              id: portrait
              description: Portrait and head studies
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid tag ID format
        '409':
          description: Tag ID already exists
...
  /api/images:
    post:
      tags:
        - Images
      summary: Upload image
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type_id
                - src
              properties:
                type_id:
                  type: integer
                src:
                  type: string
                  description: Image URL
      responses:
        '201':
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '401':
          description: Unauthorized
...